+++ test/test_approximate_time_policy.cpp
@@ -65,13 +65,13 @@ namespace message_traits
 template<>
 struct TimeStamp<Msg>
 {
-  static rclcpp::Time value(const Msg& m)
+  static rclcpp::Time value(const Msg & m)
   {
     return m.header.stamp;
   }
 };
-}
-}
+}  // namespace message_traits
+}  // namespace message_filters
 
 typedef std::pair<rclcpp::Time, rclcpp::Time> TimePair;
 typedef std::pair<rclcpp::Time, unsigned int> TimeAndTopic;
@@ -94,11 +94,11 @@ struct TimeQuad
 class ApproximateTimeSynchronizerTest
 {
 public:
-
-  ApproximateTimeSynchronizerTest(const std::vector<TimeAndTopic> &input,
-				  const std::vector<TimePair> &output,
-				  uint32_t queue_size) :
-    input_(input), output_(output), output_position_(0), sync_(queue_size)
+  ApproximateTimeSynchronizerTest(
+    const std::vector<TimeAndTopic> & input,
+    const std::vector<TimePair> & output,
+    uint32_t queue_size)
+  : input_(input), output_(output), output_position_(0), sync_(queue_size)
   {
     sync_.registerCallback(
       std::bind(
@@ -108,10 +108,8 @@ public:
         std::placeholders::_2));
   }
 
-  void callback(const MsgConstPtr& p, const MsgConstPtr& q)
+  void callback(const MsgConstPtr & p, const MsgConstPtr & q)
   {
-    //printf("Call_back called\n");
-    //printf("Call back: <%f, %f>\n", p->header.stamp.toSec(), q->header.stamp.toSec());
     ASSERT_TRUE(p);
     ASSERT_TRUE(q);
     ASSERT_LT(output_position_, output_.size());
@@ -122,28 +120,23 @@ public:
 
   void run()
   {
-    for (unsigned int i = 0; i < input_.size(); i++)
-    {
-      if (input_[i].second == 0)
-      {
+    for (size_t i = 0; i < input_.size(); i++) {
+      if (input_[i].second == 0) {
         MsgPtr p(std::make_shared<Msg>());
         p->header.stamp = input_[i].first;
         sync_.add<0>(p);
-      }
-      else
-      {
+      } else {
         MsgPtr q(std::make_shared<Msg>());
         q->header.stamp = input_[i].first;
         sync_.add<1>(q);
       }
     }
-    //printf("Done running test\n");
     EXPECT_EQ(output_.size(), output_position_);
   }
 
 private:
-  const std::vector<TimeAndTopic> &input_;
-  const std::vector<TimePair> &output_;
+  const std::vector<TimeAndTopic> & input_;
+  const std::vector<TimePair> & output_;
   unsigned int output_position_;
   typedef message_filters::Synchronizer<message_filters::sync_policies::ApproximateTime<Msg,
       Msg>> Sync2;
@@ -159,11 +152,11 @@ public:
 class ApproximateTimeSynchronizerTestQuad
 {
 public:
-
-  ApproximateTimeSynchronizerTestQuad(const std::vector<TimeAndTopic> &input,
-				      const std::vector<TimeQuad> &output,
-				      uint32_t queue_size) :
-    input_(input), output_(output), output_position_(0), sync_(queue_size)
+  ApproximateTimeSynchronizerTestQuad(
+    const std::vector<TimeAndTopic> & input,
+    const std::vector<TimeQuad> & output,
+    uint32_t queue_size)
+  : input_(input), output_(output), output_position_(0), sync_(queue_size)
   {
     sync_.registerCallback(
       std::bind(
@@ -175,10 +168,10 @@ public:
         std::placeholders::_4));
   }
 
-    void callback(const MsgConstPtr& p, const MsgConstPtr& q, const MsgConstPtr& r, const MsgConstPtr& s)
+  void callback(
+    const MsgConstPtr & p, const MsgConstPtr & q, const MsgConstPtr & r,
+    const MsgConstPtr & s)
   {
-    //printf("Call_back called\n");
-    //printf("Call back: <%f, %f>\n", p->header.stamp.toSec(), q->header.stamp.toSec());
     ASSERT_TRUE(p);
     ASSERT_TRUE(q);
     ASSERT_TRUE(r);
@@ -193,12 +186,10 @@ public:
 
   void run()
   {
-    for (unsigned int i = 0; i < input_.size(); i++)
-    {
+    for (unsigned int i = 0; i < input_.size(); i++) {
       MsgPtr p(std::make_shared<Msg>());
       p->header.stamp = input_[i].first;
-      switch (input_[i].second)
-      {
+      switch (input_[i].second) {
         case 0:
           sync_.add<0>(p);
           break;
@@ -213,13 +204,12 @@ public:
           break;
       }
     }
-    //printf("Done running test\n");
     EXPECT_EQ(output_.size(), output_position_);
   }
 
 private:
-  const std::vector<TimeAndTopic> &input_;
-  const std::vector<TimeQuad> &output_;
+  const std::vector<TimeAndTopic> & input_;
+  const std::vector<TimeQuad> & output_;
   unsigned int output_position_;
   typedef message_filters::Synchronizer<message_filters::sync_policies::ApproximateTime<Msg, Msg,
       Msg, Msg>> Sync4;
@@ -232,7 +222,8 @@ public:
 //----------------------------------------------------------
 //                   Test Suite
 //----------------------------------------------------------
-TEST(ApproxTimeSync, ExactMatch) {
+TEST(ApproxTimeSync, ExactMatch)
+{
   // Input A:  a..b..c
   // Input B:  A..B..C
   // Output:   a..b..c
@@ -243,22 +234,23 @@ TEST(ApproxTimeSync, ExactMatch) {
   rclcpp::Time t(0, 0);
   rclcpp::Duration s(1, 0);
 
-  input.push_back(TimeAndTopic(t,0));     // a
-  input.push_back(TimeAndTopic(t,1));   // A
-  input.push_back(TimeAndTopic(t+s*3,0)); // b
-  input.push_back(TimeAndTopic(t+s*3,1)); // B
-  input.push_back(TimeAndTopic(t+s*6,0)); // c
-  input.push_back(TimeAndTopic(t+s*6,1)); // C
+  input.push_back(TimeAndTopic(t, 0));      // a
+  input.push_back(TimeAndTopic(t, 1));      // A
+  input.push_back(TimeAndTopic(t + s * 3, 0));  // b
+  input.push_back(TimeAndTopic(t + s * 3, 1));  // B
+  input.push_back(TimeAndTopic(t + s * 6, 0));  // c
+  input.push_back(TimeAndTopic(t + s * 6, 1));  // C
   output.push_back(TimePair(t, t));
-  output.push_back(TimePair(t+s*3, t+s*3));
-  output.push_back(TimePair(t+s*6, t+s*6));
+  output.push_back(TimePair(t + s * 3, t + s * 3));
+  output.push_back(TimePair(t + s * 6, t + s * 6));
 
   ApproximateTimeSynchronizerTest sync_test(input, output, 10);
   sync_test.run();
 }
 
 
-TEST(ApproxTimeSync, PerfectMatch) {
+TEST(ApproxTimeSync, PerfectMatch)
+{
   // Input A:  a..b..c.
   // Input B:  .A..B..C
   // Output:   ...a..b.
@@ -269,21 +261,22 @@ TEST(ApproxTimeSync, PerfectMatch) {
   rclcpp::Time t(0, 0);
   rclcpp::Duration s(1, 0);
 
-  input.push_back(TimeAndTopic(t,0));     // a
-  input.push_back(TimeAndTopic(t+s,1));   // A
-  input.push_back(TimeAndTopic(t+s*3,0)); // b
-  input.push_back(TimeAndTopic(t+s*4,1)); // B
-  input.push_back(TimeAndTopic(t+s*6,0)); // c
-  input.push_back(TimeAndTopic(t+s*7,1)); // C
-  output.push_back(TimePair(t, t+s));
-  output.push_back(TimePair(t+s*3, t+s*4));
+  input.push_back(TimeAndTopic(t, 0));      // a
+  input.push_back(TimeAndTopic(t + s, 1));    // A
+  input.push_back(TimeAndTopic(t + s * 3, 0));  // b
+  input.push_back(TimeAndTopic(t + s * 4, 1));  // B
+  input.push_back(TimeAndTopic(t + s * 6, 0));  // c
+  input.push_back(TimeAndTopic(t + s * 7, 1));  // C
+  output.push_back(TimePair(t, t + s));
+  output.push_back(TimePair(t + s * 3, t + s * 4));
 
   ApproximateTimeSynchronizerTest sync_test(input, output, 10);
   sync_test.run();
 }
 
 
-TEST(ApproxTimeSync, ImperfectMatch) {
+TEST(ApproxTimeSync, ImperfectMatch)
+{
   // Input A:  a.xb..c.
   // Input B:  .A...B.C
   // Output:   ..a...c.
@@ -294,22 +287,23 @@ TEST(ApproxTimeSync, ImperfectMatch) {
   rclcpp::Time t(0, 0);
   rclcpp::Duration s(1, 0);
 
-  input.push_back(TimeAndTopic(t,0));     // a
-  input.push_back(TimeAndTopic(t+s,1));   // A
-  input.push_back(TimeAndTopic(t+s*2,0)); // x
-  input.push_back(TimeAndTopic(t+s*3,0)); // b
-  input.push_back(TimeAndTopic(t+s*5,1)); // B
-  input.push_back(TimeAndTopic(t+s*6,0)); // c
-  input.push_back(TimeAndTopic(t+s*7,1)); // C
-  output.push_back(TimePair(t, t+s));
-  output.push_back(TimePair(t+s*6, t+s*5));
+  input.push_back(TimeAndTopic(t, 0));      // a
+  input.push_back(TimeAndTopic(t + s, 1));    // A
+  input.push_back(TimeAndTopic(t + s * 2, 0));  // x
+  input.push_back(TimeAndTopic(t + s * 3, 0));  // b
+  input.push_back(TimeAndTopic(t + s * 5, 1));  // B
+  input.push_back(TimeAndTopic(t + s * 6, 0));  // c
+  input.push_back(TimeAndTopic(t + s * 7, 1));  // C
+  output.push_back(TimePair(t, t + s));
+  output.push_back(TimePair(t + s * 6, t + s * 5));
 
   ApproximateTimeSynchronizerTest sync_test(input, output, 10);
   sync_test.run();
 }
 
 
-TEST(ApproxTimeSync, Acceleration) {
+TEST(ApproxTimeSync, Acceleration)
+{
   // Time:     0123456789012345678
   // Input A:  a...........b....c.
   // Input B:  .......A.......B..C
@@ -321,21 +315,22 @@ TEST(ApproxTimeSync, Acceleration) {
   rclcpp::Time t(0, 0);
   rclcpp::Duration s(1, 0);
 
-  input.push_back(TimeAndTopic(t,0));      // a
-  input.push_back(TimeAndTopic(t+s*7,1));  // A
-  input.push_back(TimeAndTopic(t+s*12,0)); // b
-  input.push_back(TimeAndTopic(t+s*15,1)); // B
-  input.push_back(TimeAndTopic(t+s*17,0)); // c
-  input.push_back(TimeAndTopic(t+s*18,1)); // C
-  output.push_back(TimePair(t+s*12, t+s*7));
-  output.push_back(TimePair(t+s*17, t+s*18));
+  input.push_back(TimeAndTopic(t, 0));       // a
+  input.push_back(TimeAndTopic(t + s * 7, 1));   // A
+  input.push_back(TimeAndTopic(t + s * 12, 0));  // b
+  input.push_back(TimeAndTopic(t + s * 15, 1));  // B
+  input.push_back(TimeAndTopic(t + s * 17, 0));  // c
+  input.push_back(TimeAndTopic(t + s * 18, 1));  // C
+  output.push_back(TimePair(t + s * 12, t + s * 7));
+  output.push_back(TimePair(t + s * 17, t + s * 18));
 
   ApproximateTimeSynchronizerTest sync_test(input, output, 10);
   sync_test.run();
 }
 
 
-TEST(ApproxTimeSync, DroppedMessages) {
+TEST(ApproxTimeSync, DroppedMessages)
+{
   // Queue size 1 (too small)
   // Time:     012345678901234
   // Input A:  a...b...c.d..e.
@@ -348,18 +343,18 @@ TEST(ApproxTimeSync, DroppedMessages) {
   rclcpp::Time t(0, 0);
   rclcpp::Duration s(1, 0);
 
-  input.push_back(TimeAndTopic(t,0));     // a
-  input.push_back(TimeAndTopic(t+s,1));   // A
-  input.push_back(TimeAndTopic(t+s*3,1)); // B
-  input.push_back(TimeAndTopic(t+s*4,0)); // b
-  input.push_back(TimeAndTopic(t+s*7,1)); // C
-  input.push_back(TimeAndTopic(t+s*8,0)); // c
-  input.push_back(TimeAndTopic(t+s*10,0)); // d
-  input.push_back(TimeAndTopic(t+s*11,1)); // D
-  input.push_back(TimeAndTopic(t+s*13,0)); // e
-  input.push_back(TimeAndTopic(t+s*14,1)); // E
-  output.push_back(TimePair(t+s*4, t+s*3));
-  output.push_back(TimePair(t+s*10, t+s*11));
+  input.push_back(TimeAndTopic(t, 0));       // a
+  input.push_back(TimeAndTopic(t + s, 1));     // A
+  input.push_back(TimeAndTopic(t + s * 3, 1));   // B
+  input.push_back(TimeAndTopic(t + s * 4, 0));   // b
+  input.push_back(TimeAndTopic(t + s * 7, 1));   // C
+  input.push_back(TimeAndTopic(t + s * 8, 0));   // c
+  input.push_back(TimeAndTopic(t + s * 10, 0));  // d
+  input.push_back(TimeAndTopic(t + s * 11, 1));  // D
+  input.push_back(TimeAndTopic(t + s * 13, 0));  // e
+  input.push_back(TimeAndTopic(t + s * 14, 1));  // E
+  output.push_back(TimePair(t + s * 4, t + s * 3));
+  output.push_back(TimePair(t + s * 10, t + s * 11));
 
   ApproximateTimeSynchronizerTest sync_test(input, output, 1);
   sync_test.run();
@@ -371,17 +366,18 @@ TEST(ApproxTimeSync, DroppedMessages) {
   // Output:   ....a..b...c.d.
   //           ....A..B...C.D.
   std::vector<TimePair> output2;
-  output2.push_back(TimePair(t, t+s));
-  output2.push_back(TimePair(t+s*4, t+s*3));
-  output2.push_back(TimePair(t+s*8, t+s*7));
-  output2.push_back(TimePair(t+s*10, t+s*11));
+  output2.push_back(TimePair(t, t + s));
+  output2.push_back(TimePair(t + s * 4, t + s * 3));
+  output2.push_back(TimePair(t + s * 8, t + s * 7));
+  output2.push_back(TimePair(t + s * 10, t + s * 11));
 
   ApproximateTimeSynchronizerTest sync_test2(input, output2, 2);
   sync_test2.run();
 }
 
 
-TEST(ApproxTimeSync, LongQueue) {
+TEST(ApproxTimeSync, LongQueue)
+{
   // Queue size 5
   // Time:     012345678901234
   // Input A:  abcdefghiklmnp.
@@ -417,7 +413,8 @@ TEST(ApproxTimeSync, LongQueue) {
 }
 
 
-TEST(ApproxTimeSync, DoublePublish) {
+TEST(ApproxTimeSync, DoublePublish)
+{
   // Input A:  a..b
   // Input B:  .A.B
   // Output:   ...b
@@ -431,19 +428,20 @@ TEST(ApproxTimeSync, DoublePublish) {
   rclcpp::Time t(0, 0);
   rclcpp::Duration s(1, 0);
 
-  input.push_back(TimeAndTopic(t,0));     // a
-  input.push_back(TimeAndTopic(t+s,1));   // A
-  input.push_back(TimeAndTopic(t+rclcpp::Duration(3, 0), 1)); // B
-  input.push_back(TimeAndTopic(t+rclcpp::Duration(3, 0), 0)); // b
-  output.push_back(TimePair(t, t+s));
-  output.push_back(TimePair(t+rclcpp::Duration(3, 0), t+rclcpp::Duration(3, 0)));
+  input.push_back(TimeAndTopic(t, 0));     // a
+  input.push_back(TimeAndTopic(t + s, 1));   // A
+  input.push_back(TimeAndTopic(t + rclcpp::Duration(3, 0), 1));  // B
+  input.push_back(TimeAndTopic(t + rclcpp::Duration(3, 0), 0));  // b
+  output.push_back(TimePair(t, t + s));
+  output.push_back(TimePair(t + rclcpp::Duration(3, 0), t + rclcpp::Duration(3, 0)));
 
   ApproximateTimeSynchronizerTest sync_test(input, output, 10);
   sync_test.run();
 }
 
 
-TEST(ApproxTimeSync, FourTopics) {
+TEST(ApproxTimeSync, FourTopics)
+{
   // Time:     012345678901234
   // Input A:  a....e..i.m..n.
   // Input B:  .b....g..j....o
@@ -459,31 +457,38 @@ TEST(ApproxTimeSync, FourTopics) {
   rclcpp::Time t(0, 0);
   rclcpp::Duration s(1, 0);
 
-  input.push_back(TimeAndTopic(t,0));     // a
-  input.push_back(TimeAndTopic(t+s,1));   // b
-  input.push_back(TimeAndTopic(t+rclcpp::Duration(2, 0),2));   // c
-  input.push_back(TimeAndTopic(t+rclcpp::Duration(3, 0),3));   // d
-  input.push_back(TimeAndTopic(t+rclcpp::Duration(5, 0),0));   // e
-  input.push_back(TimeAndTopic(t+rclcpp::Duration(5, 0),3));   // f
-  input.push_back(TimeAndTopic(t+rclcpp::Duration(6, 0),1));   // g
-  input.push_back(TimeAndTopic(t+rclcpp::Duration(6, 0),2));   // h
-  input.push_back(TimeAndTopic(t+rclcpp::Duration(8, 0),0));   // i
-  input.push_back(TimeAndTopic(t+rclcpp::Duration(9, 0),1));   // j
-  input.push_back(TimeAndTopic(t+rclcpp::Duration(10, 0),2));   // k
-  input.push_back(TimeAndTopic(t+rclcpp::Duration(11, 0),3));   // l
-  input.push_back(TimeAndTopic(t+rclcpp::Duration(10, 0),0));   // m
-  input.push_back(TimeAndTopic(t+rclcpp::Duration(13, 0),0));   // n
-  input.push_back(TimeAndTopic(t+rclcpp::Duration(14, 0),1));   // o
-  output.push_back(TimeQuad(t, t+s, t+rclcpp::Duration(2, 0), t+rclcpp::Duration(3, 0)));
-  output.push_back(TimeQuad(t+rclcpp::Duration(5, 0), t+rclcpp::Duration(6, 0), t+rclcpp::Duration(6, 0), t+rclcpp::Duration(5, 0)));
-  output.push_back(TimeQuad(t+rclcpp::Duration(10, 0), t+rclcpp::Duration(9, 0), t+rclcpp::Duration(10, 0), t+rclcpp::Duration(11, 0)));
+  input.push_back(TimeAndTopic(t, 0));     // a
+  input.push_back(TimeAndTopic(t + s, 1));   // b
+  input.push_back(TimeAndTopic(t + rclcpp::Duration(2, 0), 2));   // c
+  input.push_back(TimeAndTopic(t + rclcpp::Duration(3, 0), 3));   // d
+  input.push_back(TimeAndTopic(t + rclcpp::Duration(5, 0), 0));   // e
+  input.push_back(TimeAndTopic(t + rclcpp::Duration(5, 0), 3));   // f
+  input.push_back(TimeAndTopic(t + rclcpp::Duration(6, 0), 1));   // g
+  input.push_back(TimeAndTopic(t + rclcpp::Duration(6, 0), 2));   // h
+  input.push_back(TimeAndTopic(t + rclcpp::Duration(8, 0), 0));   // i
+  input.push_back(TimeAndTopic(t + rclcpp::Duration(9, 0), 1));   // j
+  input.push_back(TimeAndTopic(t + rclcpp::Duration(10, 0), 2));   // k
+  input.push_back(TimeAndTopic(t + rclcpp::Duration(11, 0), 3));   // l
+  input.push_back(TimeAndTopic(t + rclcpp::Duration(10, 0), 0));   // m
+  input.push_back(TimeAndTopic(t + rclcpp::Duration(13, 0), 0));   // n
+  input.push_back(TimeAndTopic(t + rclcpp::Duration(14, 0), 1));   // o
+  output.push_back(TimeQuad(t, t + s, t + rclcpp::Duration(2, 0), t + rclcpp::Duration(3, 0)));
+  output.push_back(
+    TimeQuad(
+      t + rclcpp::Duration(5, 0), t + rclcpp::Duration(6, 0),
+      t + rclcpp::Duration(6, 0), t + rclcpp::Duration(5, 0)));
+  output.push_back(
+    TimeQuad(
+      t + rclcpp::Duration(10, 0), t + rclcpp::Duration(9, 0),
+      t + rclcpp::Duration(10, 0), t + rclcpp::Duration(11, 0)));
 
   ApproximateTimeSynchronizerTestQuad sync_test(input, output, 10);
   sync_test.run();
 }
 
 
-TEST(ApproxTimeSync, EarlyPublish) {
+TEST(ApproxTimeSync, EarlyPublish)
+{
   // Time:     012345678901234
   // Input A:  a......e
   // Input B:  .b......
@@ -499,19 +504,20 @@ TEST(ApproxTimeSync, EarlyPublish) {
   rclcpp::Time t(0, 0);
   rclcpp::Duration s(1, 0);
 
-  input.push_back(TimeAndTopic(t,0));     // a
-  input.push_back(TimeAndTopic(t+s,1));   // b
-  input.push_back(TimeAndTopic(t+s*2,2));   // c
-  input.push_back(TimeAndTopic(t+s*3,3));   // d
-  input.push_back(TimeAndTopic(t+s*7,0));   // e
-  output.push_back(TimeQuad(t, t+s, t+s*2, t+s*3));
+  input.push_back(TimeAndTopic(t, 0));     // a
+  input.push_back(TimeAndTopic(t + s, 1));   // b
+  input.push_back(TimeAndTopic(t + s * 2, 2));   // c
+  input.push_back(TimeAndTopic(t + s * 3, 3));   // d
+  input.push_back(TimeAndTopic(t + s * 7, 0));   // e
+  output.push_back(TimeQuad(t, t + s, t + s * 2, t + s * 3));
 
   ApproximateTimeSynchronizerTestQuad sync_test(input, output, 10);
   sync_test.run();
 }
 
 
-TEST(ApproxTimeSync, RateBound) {
+TEST(ApproxTimeSync, RateBound)
+{
   // Rate bound A: 1.5
   // Input A:  a..b..c.
   // Input B:  .A..B..C
@@ -523,17 +529,17 @@ TEST(ApproxTimeSync, RateBound) {
   rclcpp::Time t(0, 0);
   rclcpp::Duration s(1, 0);
 
-  input.push_back(TimeAndTopic(t,0));     // a
-  input.push_back(TimeAndTopic(t+s,1));   // A
-  input.push_back(TimeAndTopic(t+s*3,0)); // b
-  input.push_back(TimeAndTopic(t+s*4,1)); // B
-  input.push_back(TimeAndTopic(t+s*6,0)); // c
-  input.push_back(TimeAndTopic(t+s*7,1)); // C
-  output.push_back(TimePair(t, t+s));
-  output.push_back(TimePair(t+s*3, t+s*4));
+  input.push_back(TimeAndTopic(t, 0));      // a
+  input.push_back(TimeAndTopic(t + s, 1));    // A
+  input.push_back(TimeAndTopic(t + s * 3, 0));  // b
+  input.push_back(TimeAndTopic(t + s * 4, 1));  // B
+  input.push_back(TimeAndTopic(t + s * 6, 0));  // c
+  input.push_back(TimeAndTopic(t + s * 7, 1));  // C
+  output.push_back(TimePair(t, t + s));
+  output.push_back(TimePair(t + s * 3, t + s * 4));
 
   ApproximateTimeSynchronizerTest sync_test(input, output, 10);
-  sync_test.sync_.setInterMessageLowerBound(0, s*1.5);
+  sync_test.sync_.setInterMessageLowerBound(0, s * 1.5);
   sync_test.run();
 
   // Rate bound A: 2
@@ -542,16 +548,16 @@ TEST(ApproxTimeSync, RateBound) {
   // Output:   .a..b..c
   //           .A..B..C
 
-  output.push_back(TimePair(t+s*6, t+s*7));
+  output.push_back(TimePair(t + s * 6, t + s * 7));
 
   ApproximateTimeSynchronizerTest sync_test2(input, output, 10);
-  sync_test2.sync_.setInterMessageLowerBound(0, s*2);
+  sync_test2.sync_.setInterMessageLowerBound(0, s * 2);
   sync_test2.run();
-
 }
 
 
-int main(int argc, char **argv) {
+int main(int argc, char ** argv)
+{
   testing::InitGoogleTest(&argc, argv);
   rclcpp::init(argc, argv);
